added tests for SizeGrid & configuration for jsdom
