+++ test/dom.js
@@ -1,0 +1,21 @@
+import { JSDOM } from 'jsdom';
+ 
+const { window } = new JSDOM('<!doctype html><html><body></body></html>');
+ 
+function copyProps(src, target) {
+  const props = Object.getOwnPropertyNames(src)
+    .filter(prop => typeof target[prop] === 'undefined')
+    .reduce((result, prop) => ({
+      ...result,
+      [prop]: Object.getOwnPropertyDescriptor(src, prop),
+    }), {});
+  Object.defineProperties(target, props);
+}
+ 
+global.window = window;
+global.document = window.document;
+global.navigator = {
+  userAgent: 'node.js',
+};
+ 
+copyProps(window, global);
